hello_world/requirements.txt
Flask==1.1.2
uwsgi==2.0.19.1
gunicorn==20.1.0
google-cloud-automl
